editor/libeditor/html/tests/test_bug432225.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=432225
-->
<head>
  <title>Test for Bug 432225</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>  
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=432225">Mozilla Bug 432225</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 432225 **/

SimpleTest.waitForExplicitFinish();
addLoadEvent(runTest);

var gMisspeltWords = [];

function getEdit() {
  return document.getElementById('edit');
}

function editDoc() {
  return getEdit().contentDocument;
}

function getSpellCheckSelection() {
  var Ci = Components.interfaces;
  var win = editDoc().defaultView;
  var editingSession = win.QueryInterface(Ci.nsIInterfaceRequestor)
                          .getInterface(Ci.nsIWebNavigation)
                          .QueryInterface(Ci.nsIInterfaceRequestor)
                          .getInterface(Ci.nsIEditingSession);
  var editor = editingSession.getEditorForWindow(win);
  var selcon = editor.selectionController;
  return selcon.getSelection(selcon.SELECTION_SPELLCHECK);  
}

function runTest() {
  editDoc().designMode = "on";
  setTimeout(function() { addWords(100); }, 0);
}  
 
function addWords(aLimit) {
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
  if (aLimit == 0) {
    is(isSpellingCheckOk(), true, "All misspellings accounted for.");
    SimpleTest.finish();
    return;
  }
  getEdit().focus();
  sendChar('a', editDoc());
  sendChar('a', editDoc());
  sendChar(' ', editDoc());
  sendChar('O', editDoc());
  sendChar('K', editDoc());
  sendChar(' ', editDoc());
  gMisspeltWords.push("aa");
  setTimeout(function() { addWords(aLimit-1); }, 0);
}

function isSpellingCheckOk() {
  var sel = getSpellCheckSelection();
  var numWords = sel.rangeCount;
  
  is(numWords, gMisspeltWords.length, "Correct number of misspellings and words.");
  
  if (numWords != gMisspeltWords.length) 
    return false;
  
  for (var i=0; i<numWords; i++) {
    var word = sel.getRangeAt(i);
    is (word, gMisspeltWords[i], "Misspelling is what we think it is.");
    if (word != gMisspeltWords[i])
      return false;
  }
  return true;
}

</script>
</pre>

<iframe id="edit" width="200" height="100" src="about:blank"></iframe>

</body>
</html>
